Fix mysterious `Missing objects: ()` error and fix leaving engine in an inconsistent state when downloading objects made with unsupported extensions.
If mounting the object fails, actually make sure to propagate the reason back to the user and delete the downloaded object (so that the downloader doesn't think the object already exists). This can be wasteful and in the future we should fix the logic to somehow take into account both objects that weren't downloaded and objects that didn't have a foreign table created for them but just asking the user to redownload is good enough for now.

Also, don't try to run various cache management functions etc if no objects were successful (cause of the `object_id IN ()` issue)

+++ splitgraph/core/object_manager.py
@@ -241,7 +241,8 @@ class ObjectManager(FragmentManager):
                 partial_failure = self._generate_download_error(table, difference, cause=e)
 
             # No matter what, claim the space required by the newly downloaded objects.
-            self._increase_cache_occupancy(successful)
+            if successful:
+                self._increase_cache_occupancy(successful)
 
             if partial_failure:
                 # Instead of deleting all objects in this batch, discard the cache data
@@ -673,6 +674,8 @@ class ObjectManager(FragmentManager):
         return to_delete, freed_space
 
     def _delete_cache_entries(self, to_delete: List[str]) -> None:
+        if not to_delete:
+            return
         self.object_engine.run_sql(
             SQL(
                 "DELETE FROM {}.{} WHERE object_id IN ("
